[S07P31D206-207] feat : mission clear api

+++ BE/greenEarth/src/main/java/com/ssafy/greenEarth/repository/MissionLogRepository.java
@@ -4,5 +4,5 @@ import com.ssafy.greenEarth.domain.MissionLog;
 import org.springframework.data.jpa.repository.JpaRepository;
 
 public interface MissionLogRepository extends JpaRepository<MissionLog, Integer> {
-    MissionLog findById(Long id);
+    MissionLog findById(int id);
 }
